+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog][https://keepachangelog.com/en/1.0.0/] 
 ### Added
 
 - Report errors when there are attributes in an invalid position
+- (experimental) Add `schema` attribute
 
 ### Fixed